+++ app/components/timeline/timeline-service.js
@@ -39,7 +39,8 @@ angular.module('lizard-nxt')
   futureIndicator,
   aggWindow, // aggregation window
   lines, // events start - end
-  bars; // rain intensity
+  bars, // rain intensity
+  tickmarks; // data availability indicators
 
   /**
    * @constructor
@@ -257,9 +258,11 @@ angular.module('lizard-nxt')
      */
     updateElements: {
       value: function (oldDimensions, timestamp, interval) {
+
         if (bars && oldDimensions) {
           updateRectangleElements(bars, xScale, oldDimensions, this.dimensions);
         }
+
         if (futureIndicator) {
           updateFutureIndicator(
             futureIndicator,
@@ -268,9 +271,15 @@ angular.module('lizard-nxt')
             this.dimensions
           );
         }
+
         if (aggWindow) {
           this.drawAggWindow(timestamp, interval, oldDimensions);
         }
+
+        if (tickmarks) {
+          updateTickmarks(tickmarks, this.dimensions);
+        }
+
       }
     },
 
@@ -299,6 +308,13 @@ angular.module('lizard-nxt')
           slug,
           color
         );
+      }
+    },
+
+    drawTickMarks: {
+      value: function(data, slug) {
+        tickmarks = drawTickMarkElements(
+          this._svg, this.dimensions, data, slug);
       }
     },
 
@@ -492,6 +508,11 @@ angular.module('lizard-nxt')
       .attr('height', height)
       .attr('width', width)
       .attr('id', 'circle-group');
+    // Create group for tickmarks
+    svg.select('g').append('g')
+      .attr('height', height)
+      .attr('width', width)
+      .attr('id', 'tickmark-group');
 
     return svg;
 
@@ -598,10 +619,16 @@ angular.module('lizard-nxt')
           .attr("x1", xOneFunction)
           .attr("x2", xTwoFunction);
       }
+
+      if (tickmarks) {
+        updateTickmarks(tickmarks, dimensions);
+      }
+
       if (zoomFn) {
         zoomFn(xScale);
       }
     };
+
     return zoomed;
   };
 
@@ -720,6 +747,80 @@ angular.module('lizard-nxt')
        .duration(Timeline.prototype.transTime)
        .attr('height', height);
     }
+  };
+  var updateTickmarks = function (tickmarks, dimensions) {
+    tickmarks.attr("x", function (d) { return xScale(d); });
+  };
+
+  /**
+   * @function
+   * @summary Draws rectangular tickmarks for every timestamp in data array.
+   *
+   * @param {object} svg - timeline svg object.
+   * @param {object} dimensions - timeline dimensions object.
+   * @param {integer []} data - list of timestamps in ms.
+   *
+   * @returns {object} d3 selection object with tickmarks for each timestamp.
+   */
+  var drawTickMarkElements = function (svg, dimensions, data, slug) {
+    var height = Timeline.prototype._getHeight(dimensions);
+
+    // setup svg group element to hold rects.
+    if (data.length > 0) {
+
+      var group = svg.select("g").select("#tickmark-group").select("#" + slug);
+
+      // if group element doesn't exist yet, create one
+      if (!group[0][0]) {
+        group = svg.select("g").select("#tickmark-group").append("g")
+                      .attr("id", slug);
+      }
+
+      // DATA JOIN
+      // Join new data with old elements, based on the timestamp.
+      tickmarks = group.selectAll("rect")
+        .data(data, function  (d) { return d; });
+
+    } else if (data.length === 0 && slug === undefined) {
+      // if no data and slugs are defined, remove all groups
+      svg.select("g").select("#tickmark-group").selectAll("g")
+        .remove();
+
+      return;
+    }
+
+    // UPDATE
+    tickmarks.transition()
+      .delay(Timeline.prototype.transTime)
+      .duration(Timeline.prototype.transTime)
+      .attr("y", height - 5)
+      .attr("x", function (d) { return xScale(d); } );
+
+    // ENTER
+    tickmarks.enter().append("rect")
+      .attr("class", "tickmark")
+    .transition()
+      .delay(Timeline.prototype.transTime)
+      .duration(Timeline.prototype.transTime)
+      .attr("x", function (d) { return xScale(d); } )
+      .attr("y", height - 5)
+      .attr("height", 5)
+      .attr("width", 2);
+
+    // EXIT
+    // Remove old elements as needed.
+    tickmarks.exit()
+      .transition()
+      .delay(0)
+      .duration(Timeline.prototype.transTime)
+    .transition()
+      .delay(Timeline.prototype.transTime)
+      .duration(Timeline.prototype.transTime)
+      .attr("stroke-width", 0)
+      .remove();
+
+    return tickmarks;
+
   };
 
   /**